fixup setup.py (only write project-config.jam unless --project-config is specified)

+++ bindings/python/setup.py
@@ -409,20 +409,22 @@ class LibtorrentBuildExt(BuildExtBase):
 
         # We use a "project-config.jam" to instantiate a python environment
         # to exactly match the running one.
-        config = open(python_binding_dir / 'project-config.jam', 'w+')
         try:
-            write_b2_python_config(config)
-            config.seek(0)
-            log.info("project-config.jam contents:")
-            log.info(config.read())
-            config.close()
+            if self._should_add_arg("--project-config"):
+                config = open(python_binding_dir / 'project-config.jam', 'w+')
+                write_b2_python_config(config)
+                config.seek(0)
+                log.info("project-config.jam contents:")
+                log.info(config.read())
+                config.close()
             yield
 
         finally:
             # If we errored while writing config, windows may complain about
             # unlinking a file "in use"
-            config.close()
-            os.unlink(python_binding_dir / 'project-config.jam')
+            if self._should_add_arg("--project-config"):
+                config.close()
+                os.unlink(python_binding_dir / 'project-config.jam')
 
 
 setuptools.setup(
